improve unit tests on scenarios
